+++ tor/core/mentions.py
@@ -1,7 +1,7 @@
 import logging
 
 from tor.core.helpers import _
-from tor.strings.responses import pm_body, pm_subject
+from tor.strings import translation
 
 
 def process_mention(mention):
@@ -12,6 +12,9 @@ def process_mention(mention):
     :param mention: the Comment object containing the username mention.
     :return: None.
     """
+    i18n = translation()
+    pm_subject = i18n['responses']['direct_message']['subject']
+    pm_body = i18n['responses']['direct_message']['body']
 
     # message format is subject, then body
     mention.author.message(pm_subject, _(pm_body))