+++ tor/strings/urls.py
@@ -1,3 +1,0 @@
-yt_transcript_url = 'http://video.google.com/timedtext?lang=en&v={}'
-
-ToR_link = 'www.reddit.com/r/TranscribersOfReddit'